DesktopAgent/requirements.txt
websockets>=11.0.3
aiohttp>=3.8.0
google-cloud-logging>=3.0.0
